Refactor Bounty Page for enterprise: Bounty form now accepts a variant: edit or create. Depending on this option, the form will display the options to create a form and, if it's edit, the options to edit a form. This makes the form reusable depending on the need
